requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.3.4
streamlit==1.8.1
openpyxl==3.0.9
protobuf==3.20.*
XlsxWriter==3.0.1
gspread==5.7.2
google-auth-httplib2==0.1.0